visuals_created & time_added

+++ src/config.py
@@ -6,6 +6,6 @@ world shaped with 21x21 size
 
 """
 
-TERRITORY_SIZE: float = 53
+TERRITORY_SIZE: int = 53
 PLUME_SIZE: float = 5
 WIND_DIRECTION: str = "WEST"
